api/requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.0
google-cloud-bigquery==3.27.0
google-auth==2.38.0
openai==1.73.0
pydantic==2.11.1
python-dotenv==1.1.0
